+++ scripts/build-netbsd-pkg.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 # SPDX-License-Identifier: GPL-3.0-or-later
-# Build a NetBSD binary package (.tgz) for merovingian 0.10.3.
+# Build a NetBSD binary package (.tgz) for merovingian 0.10.4.
 #
 # The checked-in packaging/netbsd/Makefile is the pkgsrc recipe kept for
 # downstream porters.  This script generates a framework-free package so CI
@@ -11,7 +11,7 @@
 # a staged tree using tar, which produces a package that pkg_add can install.
 set -e
 
-VERSION="0.10.3"
+VERSION="0.10.4"
 STAGE="staging-netbsd"
 PREFIX=/usr/pkg
 
@@ -72,6 +72,7 @@ cat > "+CONTENTS" <<EOF
 bin/merovingian-db-migrate
 bin/merovingian-server
 etc/merovingian/merovingian.conf.sample
+libexec/merovingian/merovingian-fed-worker
 libexec/merovingian/merovingian-thumbnail-worker
 share/examples/rc.d/merovingian
 @dir share/examples/rc.d
@@ -95,6 +96,8 @@ cp "${STAGE}${PREFIX}/bin/merovingian-db-migrate"        "${PKGDIR}/bin/"
 cp "${STAGE}${PREFIX}/bin/merovingian-server"            "${PKGDIR}/bin/"
 cp "${STAGE}${PREFIX}/etc/merovingian/merovingian.conf.sample" \
                                                           "${PKGDIR}/etc/merovingian/"
+cp "${STAGE}${PREFIX}/libexec/merovingian/merovingian-fed-worker" \
+                                                          "${PKGDIR}/libexec/merovingian/"
 cp "${STAGE}${PREFIX}/libexec/merovingian/merovingian-thumbnail-worker" \
                                                           "${PKGDIR}/libexec/merovingian/"
 cp "${STAGE}${PREFIX}/share/examples/rc.d/merovingian"   "${PKGDIR}/share/examples/rc.d/"
@@ -104,6 +107,7 @@ cp "${STAGE}${PREFIX}/share/examples/rc.d/merovingian"   "${PKGDIR}/share/exampl
     bin/merovingian-db-migrate \
     bin/merovingian-server \
     etc/merovingian/merovingian.conf.sample \
+    libexec/merovingian/merovingian-fed-worker \
     libexec/merovingian/merovingian-thumbnail-worker \
     share/examples/rc.d/merovingian )